[ front / feat ] : 유저 생성시에 기본 이미지 커밋 에러 해결
