+++ backend/src/main/java/com/example/booktree/LikePost/dto/response/LikePostResponseDto.java
@@ -1,4 +1,4 @@
-package com.example.booktree.LikePost.dto.response;
+package com.example.booktree.likepost.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
