+++ backend/src/main/java/com/example/booktree/likecomment/dto/LikeCommentDto.java
@@ -1,4 +1,4 @@
-package com.example.booktree.like_reply.dto;
+package com.example.booktree.likecomment.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -6,23 +6,22 @@ import lombok.Getter;
 import lombok.Setter;
 
 @Data
-public class LikeReplyDto {
+public class LikeCommentDto {
 
     @Getter
     @Setter
     @AllArgsConstructor
     public static class Post {
-        // 좋아요할 대댓글의 ID
-        private Long replyId;
+        private Long commentId;
     }
 
+    // 응답용
     @Getter
     @Setter
     @AllArgsConstructor
     public static class Response {
-        private Long likeReplyId;
-        private Long replyId; // 좋아요 대상 대댓글의 ID
+        private Long likeCommentId;
+        private Long commentId; // 좋아요 대상 댓글의 ID
         private Long userId; // 좋아요를 누른 사용자의 ID
-        private Long likeCount;
     }
 }
