+++ backend/src/main/java/com/example/booktree/popularpost/service/PopularPostService.java
@@ -1,6 +1,8 @@
 
 package com.example.booktree.popularpost.service;
 
+
+import static com.example.booktree.utils.ImageUtil.DEFAULT_POST_IMAGE;
 import com.example.booktree.exception.BusinessLogicException;
 import com.example.booktree.exception.ExceptionCode;
 import com.example.booktree.post.dto.response.PostResponseDto;
@@ -30,7 +32,7 @@ public class PopularPostService {
     private final StringRedisTemplate redisTemplate;
     private final PostService postService;
     private static final String REDIS_KEY = "popular:posts:";
-    private final String defaultImageUrl = "https://booktree-s3-bucket.s3.ap-northeast-2.amazonaws.com/default_profile.png";
+    private final String defaultImageUrl = DEFAULT_POST_IMAGE;
 
 
     // 메인 카테고리 별 게시글 조회 시 인기순위에 반영
