+++ backend/src/main/java/com/example/booktree/reply/entity/Reply.java
@@ -3,7 +3,7 @@ package com.example.booktree.reply.entity;
 
 import com.example.booktree.auditable.Auditable;
 import com.example.booktree.comment.entity.Comment;
-import com.example.booktree.like_reply.entity.LikeReply;
+import com.example.booktree.likereply.entity.LikeReply;
 import com.example.booktree.user.entity.User;
 import jakarta.persistence.*;
 import java.util.ArrayList;
