+++ backend/src/main/java/com/example/booktree/user/entity/User.java
@@ -4,8 +4,8 @@ package com.example.booktree.user.entity;
 import com.example.booktree.auditable.Auditable;
 import com.example.booktree.blog.entity.Blog;
 import com.example.booktree.category.entity.Category;
-import com.example.booktree.like_comment.entity.LikeComment;
-import com.example.booktree.like_reply.entity.LikeReply;
+import com.example.booktree.likecomment.entity.LikeComment;
+import com.example.booktree.likereply.entity.LikeReply;
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -13,7 +13,6 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
-import jakarta.websocket.Encoder.Text;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
