--- backend/src/main/java/com/example/booktree/utils/ConstData.java
@@ -1,8 +1,0 @@
-package com.example.booktree.utils;
-
-public class ConstData {
-
-    private static String DEFAULT_IMAGE = "https://booktree-s3-bucket.s3.ap-northeast-2.amazonaws.com/default_profile.png";
-
-
-}
