+++ frontend/src/app/mypage/[id]/page.tsx
@@ -281,7 +281,7 @@ export default function MyPage() {
                             <img
                                 src={
                                     loginUser.image ||
-                                    'https://booktree-s3-bucket.s3.ap-northeast-2.amazonaws.com/default_profile.png'
+                                    'https://booktree-s3-bucket.s3.ap-northeast-2.amazonaws.com/%E1%84%89%E1%85%A1%E1%84%8B%E1%85%AD%E1%86%BC%E1%84%8C%E1%85%A1%E1%84%80%E1%85%B5%E1%84%87%E1%85%A9%E1%86%AB%E1%84%8B%E1%85%B5%E1%84%86%E1%85%B5%E1%84%8C%E1%85%B5.png'
                                 }
                                 alt="프로필"
                                 className="w-full h-full object-cover"
