+++ terraform.tfstate
@@ -1,0 +1,9 @@
+{
+  "version": 4,
+  "terraform_version": "1.11.4",
+  "serial": 1,
+  "lineage": "7ddcf2af-d52d-4f62-7a73-2cdcf037d813",
+  "outputs": {},
+  "resources": [],
+  "check_results": null
+}
